cli/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "xcoin-cli"
version = "0.1.0"
description = "xcoinalgo CLI - Strategy development toolkit for quant teams"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "xcoinalgo", email = "dev@xcoinalgo.com"}
]
keywords = ["trading", "algorithmic-trading", "cryptocurrency", "strategy", "cli"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]

dependencies = [
    "click>=8.0.0",
    "rich>=13.0.0",
    "requests>=2.28.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "PyYAML>=6.0",
    "python-dotenv>=1.0.0",
    "cryptography>=41.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]

[project.scripts]
xcoin = "xcoin_cli.cli:cli"

[project.urls]
Homepage = "https://xcoinalgo.com"
Documentation = "https://docs.xcoinalgo.com/cli"
Repository = "https://github.com/xcoinalgo/xcoin-cli"
Issues = "https://github.com/xcoinalgo/xcoin-cli/issues"

[tool.setuptools]
packages = ["xcoin_cli", "xcoin_cli.commands", "xcoin_cli.templates"]

[tool.setuptools.package-data]
xcoin_cli = ["templates/*"]

[tool.black]
line-length = 100
target-version = ["py38", "py39", "py310", "py311"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "--cov=xcoin_cli --cov-report=html --cov-report=term-missing"

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
